requirements.txt
flask==2.1.2
flask_oauthlib
sqlalchemy==1.4.39
pymysql==1.0.2
requests
flask-cors
boto3
pyjwt